requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
efficientnet-pytorch==0.7.1
eval_type_backport==0.2.2
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.2
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pretrainedmodels==0.7.4
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.13.1
segmentation_models_pytorch==0.4.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.2
sympy==1.13.1
timm==1.0.15
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
open3d

